Add alert when tapping icon and make SiteIconPicker not a VC
